Renaming alt and config provider to camel case, taking a step back from making the review page a client component and instead will be leveraging CSS
