+++ src/app/privacy-policy/page.tsx
@@ -1,7 +1,7 @@
 import { getDocument } from "../../db/db";
 import Text from "@/components/text/text";
 import Article from "@/components/article/article";
-import type { Config, Text as TextType } from "@/lib/config-provider";
+import type { Config, Text as TextType } from "@/lib/configProvider";
 
 import "./privacy_policy.css";
