+++ src/app/reviews/[slug]/page.tsx
@@ -1,11 +1,14 @@
 import { notFound } from "next/navigation";
 import { Review } from "./review";
 import getCompanyData from "@/lib/getCompanyData";
-import type { Company } from "../columns";
 import { isValidSlug } from "@/lib/utils";
 import Article from "@/components/article/article";
-import { AltRecord, getAltObj } from "@/lib/altprovider";
+import { AltRecord, getAltObj, getAltString } from "@/lib/altProvider";
+import type { Company } from "../columns";
+
 export const dynamic = "force-dynamic";
+
+import "./review.css";
 
 export default async function Page({
   params,
@@ -23,6 +26,12 @@ export default async function Page({
 
   const [company, alt] = await Promise.all([companyPromise, altPromise]);
 
-  return <Review data={company} altObj={alt} />;
+  const altObj = ["review_count", "average_rating", "adjusted_review_count", "adjusted_average_rating"].reduce((obj, curr) => ({...obj, [curr]: getAltString(alt, curr, company[curr])}), {})
+
+  return (
+    <Article className="container mx-auto py-10 review-summary">
+      <Review data={company} altObj={altObj} />;
+    </Article>
+  );
 }
