+++ src/app/reviews/[slug]/review.css
@@ -27,9 +27,3 @@
         line-height: 1.25rem;
     }
 }
-
-@media screen and (max-width: 375px) {
-    #review-data-list {
-        display: none;
-    }
-}
