+++ src/lib/altProvider.ts
@@ -36,7 +36,7 @@ export async function getAltStringAsync(value: string | number, page: string, ke
 export function getAltString(altObj: AltRecord, key: string, value: any){
   if (altObj[key]) {
     const { prefix, postfix } = altObj[key];
-    return `${prefix} ${value} ${postfix}`;
+    return `${prefix} ${value} ${postfix}`.trim();
   }
   return "undefined";
 }
